Исправил не рабочую пагинацию в приютах и юзерах, написал тесты

+++ src/main/java/ru/pet/nursery/repository/AnimalRepo.java
@@ -19,10 +19,10 @@ public interface AnimalRepo extends JpaRepository<Animal, Integer> {
     @Modifying
     @Transactional
     @Query(value = "UPDATE animal_table SET telegram_user_id = :adoptedId, took_date = :date WHERE id = :id", nativeQuery = true)
-    void updateWhoTookPetAndTookDate(Long adoptedId, LocalDate date, int id);
+    void updateWhoTookPetAndTookDate(Long adoptedId, LocalDate date, Long id);
 
     @Modifying
     @Transactional
     @Query(value = "UPDATE animal_table SET telegram_user_id = 1, took_date = null, pet_return_date = :date WHERE id = :id", nativeQuery = true)
-    void updateReturnDateAnimal(LocalDate date, int id);
+    void updateReturnDateAnimal(LocalDate date, Long id);
 }
